Use root hints for IP-to-name mapping of root
Use root hints and historical root IP addresses for IP-to-name mapping
of root.  Sometimes a domain was analyzed with old hints files, and when
translating the IP address to name, there is no way to match the
obsolete IP address with a name.  Using the root hints and the
historical root addresses helps with this issue.