Recreate EDNS if options or payload are changed
dnspython >= 2.0.0 introduces an OPT class, instances of which include
the edns, ednsflags, and payload attributes.  However, the payload
attribute is immutable because it is actually the TTL value of the OPT
record.  This commit changes the code to use use_edns() instead of
trying to set payload directly.  Additionally, the EDNS options in
dnspython >= 2.0.0 are implemented as a tuple, so use_edns() is to
modify the option as well. Fixes #74.