Use high id values for nonexistent DNSKEYs
Node strings for nonexistent DNSKEYs used id=0.  This didn't used to be
a problem because it was only ever RRSIGs and DS records that produced
nonexistent DNSKEYs; there was never a problem with collisions because
those only required keytag.  However, with the introduction of CDNSKEYs,
it is possible that a nonexistent DNSKEYs is placed on the graph with an
existing DNSKEY of the same key tag.