Base analysis class on each analysis
Rather than using a hard-coded class (OfflineDomainNameAnalysis or
TTLAgnosticOfflineDomainNameAnalysis) to deserialize an analysis, base
it on the type of each analysis (recursive or authoritative).  This, in
turn, determines whether TTL is considered when aggregating RRsets for a
given analysis.

This also fixes #139, which symptoms were caused by hard-coded classes
being used inconsistently across the different commands.